refactor(ReadComicOnline): rename impl directories to show provider intent
