+++ src/ReadComicOnline/implementations/discover-section-providing/parsers.ts
