+++ src/ReadComicOnline/implementations/manga-details-providing/main.ts
