+++ src/ReadComicOnline/implementations/manga-details-providing/parsers.ts
