+++ src/ReadComicOnline/implementations/search-result-providing/parsers.ts
