+++ src/ReadComicOnline/main.ts
@@ -1,10 +1,10 @@
 import type { Cookie, Extension, MangaProviding } from "@paperback/types";
 import { BasicRateLimiter, CookieStorageInterceptor } from "@paperback/types";
 import { ChapterProvider } from "./implementations/chapter-providing/main";
-import { DiscoverProvider } from "./implementations/discover-section/main";
-import { MangaProvider } from "./implementations/manga/main";
-import { SearchProvider } from "./implementations/search-results/main";
-import { SettingsFormProvider } from "./implementations/settings-form/forms/main";
+import { DiscoverProvider } from "./implementations/discover-section-providing/main";
+import { MangaProvider } from "./implementations/manga-details-providing/main";
+import { SearchProvider } from "./implementations/search-result-providing/main";
+import { SettingsFormProvider } from "./implementations/settings-form-providing/forms/main";
 import { applyMixins } from "./implementations/shared/utils";
 import { ReadComicOnlineInterceptor } from "./services/network";
